Can create project into folder in CreateNewChatbox Page
Code from visual studio connects to revit app data to create a new project & create local folder to save/store chatbox txt files which should be appendable, as in each time a chatbox is created within said specific project it should save the previous file and create a new appended file. As in if on CreateNewChat.html, when the user clicks Create New Project and puts in a Name & Description, once they click Done, this will create a new project inside Revit that the user can interact within, it will also at the same time open a chat box to allow the user to start interacting with the AI. When the user clicks Import Existing Project, this will showcase inside PreviousProjects_scroll a log of all Revit Projects the user has previously done, when a user clicks one, it will then allow user to write a Name & Description for project so edit, so that data can be stored withing the app data local folder, where all the chatbox appended txt files will be stored.