Allows Delete, CreateNewChat & Sample Project in Scrollbox
Essentially had to work in index.js for these three to work together. Delete - deletes a project file. CreateNewChat - within ExistingProjectPage, you'll be able to create a chatbox and go onto it.  In CreateNewChat, the scrollbox takes the projects in the Samples folder and puts them into scrollbox to be seen.